Product except self

+++ src/Prefix_Sum/FindPrefixScore.java
@@ -5,8 +5,8 @@ public class FindPrefixScore {
         int n = nums.length;
         long[] result = new long[n];
 
-        int max = 0;
-        int sum = 0;
+        long max = 0;
+        long sum = 0;
 
         for(int i=0;  i<n; i++){
             max = Math.max(max,nums[i]);
